[APM] Use same colors for HTTP status everywhere
Use the same colors on the RPM/TPM graphs as in the status code badge.

Fixes #47817.
