+++ x-pack/legacy/plugins/apm/public/components/shared/Summary/HttpStatusBadge/index.tsx
@@ -7,27 +7,8 @@
 import React from 'react';
 import { i18n } from '@kbn/i18n';
 import { EuiToolTip, EuiBadge } from '@elastic/eui';
-import theme from '@elastic/eui/dist/eui_theme_light.json';
 import { statusCodes } from './statusCodes';
-
-const {
-  euiColorDarkShade,
-  euiColorSecondary,
-  euiColorWarning,
-  euiColorDanger
-} = theme;
-
-function getStatusColor(status: number) {
-  const colors: { [key: string]: string } = {
-    1: euiColorDarkShade,
-    2: euiColorSecondary,
-    3: euiColorDarkShade,
-    4: euiColorWarning,
-    5: euiColorDanger
-  };
-
-  return colors[status.toString().substr(0, 1)] || 'default';
-}
+import { httpStatusCodeToColor } from '../../../../utils/httpStatusCodeToColor';
 
 interface HttpStatusBadgeProps {
   status: number;
@@ -39,7 +20,7 @@ export function HttpStatusBadge({ status }: HttpStatusBadgeProps) {
 
   return (
     <EuiToolTip content={label}>
-      <EuiBadge color={getStatusColor(status)}>
+      <EuiBadge color={httpStatusCodeToColor(status) || 'default'}>
         {status} {statusCodes[status.toString()]}
       </EuiBadge>
     </EuiToolTip>
