+++ x-pack/legacy/plugins/apm/public/selectors/chartSelectors.ts
@@ -20,6 +20,7 @@ import {
 import { asDecimal, asMillis, tpmUnit } from '../utils/formatters';
 import { IUrlParams } from '../context/UrlParamsContext/types';
 import { getEmptySeries } from '../components/shared/charts/CustomPlot/getEmptySeries';
+import { httpStatusCodeToColor } from '../utils/httpStatusCodeToColor';
 
 export interface ITpmBucket {
   title: string;
@@ -181,21 +182,16 @@ export function getTpmSeries(
 
 function colorMatch(key: string) {
   if (/ok|success/i.test(key)) {
-    return theme.euiColorVis0;
+    return theme.euiColorSecondary;
   } else if (/error|fail/i.test(key)) {
-    return theme.euiColorVis2;
+    return theme.euiColorDanger;
   }
 }
 
 function getColorByKey(keys: string[]) {
-  const assignedColors: StringMap<string> = {
-    'HTTP 2xx': theme.euiColorVis0,
-    'HTTP 3xx': theme.euiColorVis5,
-    'HTTP 4xx': theme.euiColorVis7,
-    'HTTP 5xx': theme.euiColorVis2
-  };
+  const assignedColors = ['HTTP 2xx', 'HTTP 3xx', 'HTTP 4xx', 'HTTP 5xx'];
 
-  const unknownKeys = difference(keys, Object.keys(assignedColors));
+  const unknownKeys = difference(keys, assignedColors);
   const unassignedColors: StringMap<string> = zipObject(unknownKeys, [
     theme.euiColorVis1,
     theme.euiColorVis3,
@@ -206,5 +202,5 @@ function getColorByKey(keys: string[]) {
   ]);
 
   return (key: string) =>
-    colorMatch(key) || assignedColors[key] || unassignedColors[key];
+    colorMatch(key) || httpStatusCodeToColor(key) || unassignedColors[key];
 }
